solve timezone issues by setting the right ENV var
I could use a gem with all the timezone info and I could use
active_support to display times in different time zones, but it appears
that Time itself can already display times correctly for the timezone
that your computer is in. If it's set to UTC, like many servers are,
then you should simply override it.